Finish the maintainability backlog: extractions, registry collapse, docs, page tests
C5 — two components out of the tool modules they were embedded in
-----------------------------------------------------------------
`BufferedUI`, a complete `UIProtocol` implementation of 125 lines, lived inside
`llm/tool/delegate.py`. Moves to `llm/ui/buffered_ui.py`, exported from
`zrb.llm.ui`; delegate.py 509 -> 387. Its 13 tests move to
`test/llm/ui/test_buffered_ui.py` — they patched
`zrb.llm.tool.delegate.agent_activity_registry`, which after the move would
have spied on nothing, so the targets follow the source.

`_StreamCapture`, 145 lines of bounded capture with spill-to-disk, lived inside
`llm/tool/shell.py` and could not be exercised without driving a subprocess.
Moves to `llm/tool/stream_capture.py` as `StreamCapture` — a name imported
across modules is not module-private (ADR-0035). shell.py 620 -> 496. One test
monkeypatched `shell_mod.zrb_print` to watch the console echo; the echo moved
with the class, so that patch is repointed too.

C3 — one registry shape, written three times
--------------------------------------------
Live, system and project context each carried their own copy of "scan for this
name, replace or append" and of "call each provider, swallow what it raises,
keep the non-empty results". The second is the one worth having once: these are
downstream extension points and one bad plugin must not cost the whole prompt.
Now one `_ProviderRegistry`. `_section_providers` stays separate — a dict
holding one provider per section composes by selecting, not concatenating.

C7 — the web page layer had no tests
------------------------------------
Only API routes and login/logout were covered, so home, chat and the 404
handler were unexercised: a renamed template failed at request time with
nothing to catch it. The alias tests matter because all three URLs are stacked
decorators on one handler. The 404 test asserts the `/api` branch rather than
the status code — "a missing route 404s" passes with the handler unregistered.

B7 — an API reference
---------------------
`./zrb-api-doc.sh` renders the public API from docstrings into the gitignored
`dist/api`. pdoc, not mkdocstrings: docs/ is plain markdown with no mkdocs.yml.
Output is not committed because it regenerates from source. Its `ToolCallPart`
warnings are correct and left alone — those aliases reference pydantic-ai types
deliberately held behind TYPE_CHECKING, and silencing a cosmetic warning would
cost startup time.

B8 — three real shims, four mislabelled ones
--------------------------------------------
Removed: `create_summarizer_history_processor(agent=)` (declared, never read)
and `(token_threshold=)` (a shadow of its named twin); `*names` on
`_file_section_middleware`, which existed to emit the retired `mandate`
alongside `workflow`; a `get_prompt("mandate")` docstring example.

Relabelled instead of deleted, because the "backward compatibility" label was
wrong: `HookManager.execute_hooks_simple` is a thirty-line adapter flattening
typed results into HookResult's Claude-format mapping, and the two
`input_queue` properties are the public read seam for queue state — without
them a caller reaches for the private attribute. `render_diff(ui=)` was already
gone.

Full suite 4378 passed / 21 skipped; pyright, flake8 -F and C901 clean.
