+++ docs/changelog-v2/2.58.0.md
@@ -21,6 +21,13 @@
 - **New: 18 names exported from `zrb`** (`src/zrb/__init__.py`): the managers were exported without the types needed to call them — `hook_manager` with no `HookResult` is a registry you cannot return from. Added: `Skill`, `SubAgentDefinition`, `HookEvent`, `HookContext`, `HookResult`, `PromptManager`, `Preset`, `register_preset`, `register_model_profile`, `model_capabilities`, `register_theme`, `register_always_auto_approve`, `PermissionPolicy`, `Rule`, `ALLOW`, `DENY`, `ASK`, `StrListAttr`. No `zrb/llm/__init__.py` was added on purpose: `zrb.llm` is an implicit namespace package whose eleven children each curate their own `__all__`, and giving it an eager root would make a narrow `zrb.llm.sandbox` import pull the whole LLM stack. The curated tier is `zrb` itself.
 - **Refactor: `llm/ui` no longer imports `LLMTask`, and seven lazy imports hoisted** (`src/zrb/llm/ui/*.py`, `src/zrb/llm/common_tools.py`, `src/zrb/llm/tool/plan.py`, `src/zrb/llm/prompt/live_context.py`): 27 in-function imports carried `# lazy: circular`, and the recurring hub was one edge appearing in four differently-worded chains — `profile → prompt → ui → llm_task → here`, `tool → ui → llm_task → here`, and two more. The cause turned out to be entirely *typing*: ten `ui` modules named `LLMTask`, six of them eagerly, and every single use was an annotation. The one runtime reference was a `cast()`. No protocol was extracted — that would be an abstraction with one implementation for a problem `from __future__ import annotations` and a `TYPE_CHECKING` block already solve. `zrb.llm.ui.base.ui` now reaches `zrb.llm.task` not at all, which let four cycle-named lazy imports hoist to module scope. Three more went with them: the comment on the `permission.state` trio claimed it "transitively imports zrb.llm.ui", and a static closure of its imports shows it reaches `ui` not at all — stale since some earlier refactor. 27 → 20.
 
+- **Refactor: two whole components moved out of the tool modules they were embedded in** (`src/zrb/llm/ui/buffered_ui.py`, `src/zrb/llm/tool/stream_capture.py`): `BufferedUI` — a complete `UIProtocol` implementation, 125 lines — lived inside `llm/tool/delegate.py`, which is both a layering inversion and an import edge pointing the wrong way; it is now `llm/ui/buffered_ui.py` and exported from `zrb.llm.ui`, taking `delegate.py` from 509 lines to 387. `_StreamCapture` — 145 lines of bounded capture with spill-to-disk — lived inside `llm/tool/shell.py` and could not be exercised without driving a subprocess; it is now `StreamCapture` in `llm/tool/stream_capture.py`, losing the underscore because a name imported across modules is not module-private (ADR-0035). `shell.py` falls 620 → 496. The 13 `BufferedUI` tests moved to `test/llm/ui/test_buffered_ui.py` with their patch targets repointed at the source, which is also what the >500-line split rule asks for.
+- **Refactor: `PromptManager` had one registry shape written three times** (`src/zrb/llm/prompt/manager.py`): live context, system context and project context each carried their own copy of "scan the list for this name, replace or append" and of "call each provider, swallow whatever it raises, keep the non-empty results". The second is the one worth having once — these are downstream extension points, and one bad plugin must not cost the whole prompt. Collapsed into `_ProviderRegistry`, used three times. `_section_providers` is deliberately not merged in: it is a dict holding *one* provider per section, so composing it means selecting an entry rather than concatenating all of them.
+
+- **New: `./zrb-api-doc.sh` renders the API reference from docstrings** (`zrb-api-doc.sh`, `pyproject.toml`, `docs/advanced-topics/maintainer-guide.md`): `docs/` is 175 files of which 76 are ADRs and 61 are changelog, and the only thing titled "API Reference" was a hand-written page for `LLMChatTask` — nothing covered `Task`, `CmdTask`, `Group`, inputs or envs. `pdoc` rather than `mkdocstrings`, because `docs/` is plain markdown with no `mkdocs.yml` and adopting mkdocs to render one reference is a bigger commitment than the reference warrants. Output goes to the gitignored `dist/api`: it regenerates from source, so a committed copy is only ever a stale second answer. It reads the annotations `py.typed` now exposes, and the two existing doc gates (`test_public_api_docs.py`, `test_public_api_contract.py`) are what keep the generated page free of blanks. The `ToolCallPart` parse warnings it emits are correct and left alone — `ResponseHandler`/`ToolPolicy`/`ArgumentFormatter` reference pydantic-ai types held behind `TYPE_CHECKING` on purpose, and hoisting the import to silence a cosmetic warning would trade startup time for it.
+- **New: the web page layer has smoke tests** (`test/runner/web_route/test_page_routes.py`): `test/runner/web_route/` held only API-route tests plus login/logout, so the home, chat and 404 handlers were unexercised — a template renamed on one side of `get_template("home_page/view.html")` failed at request time with nothing to catch it. The home and chat tests assert every URL alias, since all three are stacked decorators on one handler and a dropped line leaves a 404 no other test would see. The 404 test asserts the `/api` *branch* rather than the status code, because "a missing route 404s" passes with the handler not registered at all.
+- **Fix: four shims that were mislabelled, and three that were real** (`src/zrb/llm/summarizer/history_summarizer.py`, `src/zrb/llm/prompt/manager.py`, `src/zrb/llm/prompt/prompt.py`, `src/zrb/llm/hook/manager.py`, `src/zrb/llm/ui/polling_ui.py`, `src/zrb/llm/ui/event_driven_ui.py`): `create_summarizer_history_processor` carried `agent=` (declared, never read) and `token_threshold=` (a shadow of `conversational_token_threshold`) under a "# Backward compatibility" banner; both are gone. `_file_section_middleware` took `*names` to emit `mandate` and `workflow` at one position for configs predating that split — `mandate` is retired and the one caller passes one name, so it takes a single `name`. `get_prompt`'s docstring still used `get_prompt("mandate")` as its worked example. Three others turned out **not** to be shims despite saying so, and were relabelled rather than deleted: `HookManager.execute_hooks_simple` is a thirty-line adapter flattening typed results into `HookResult`'s Claude-format mapping, and `PollingUI.input_queue` / `EventDrivenUI.input_queue` are the public read seam for queue state — without them a caller has to reach for the private attribute. `render_diff(ui=)` was already gone.
+
 - **Documentation: the four highest-traffic public symbols had no docstring at all** (`src/zrb/group/group.py`, `src/zrb/task/make_task.py`, `src/zrb/task/task.py`, `src/zrb/input/str_input.py`): `Group` had no module, class or member docs across twelve public members; `make_task` had none despite 22 parameters and headline billing; `Task` and `StrInput` were `class X(Base): pass` with `__doc__ is None`, and they are the first names in every tutorial. All four now carry worked examples. `Group.add_group` is also correctly typed `AnyGroup | str` — it has always accepted a bare string as shorthand — and the alias-then-name fallback in `remove_group`/`remove_task` is documented rather than left to be discovered.
 
 ### Also in this release (previously staged as 2.57.1, never published)
