+++ pyproject.toml
@@ -193,6 +193,7 @@ pytest-asyncio = "^1.3.0"
 tomlkit = "^0.13.3"
 httpx2 = ">=2.4.0"
 radon = "^6.0.1"
+pdoc = ">=15"
 
 [tool.pytest.ini_options]
 asyncio_default_fixture_loop_scope = "function"
