+++ src/zrb/llm/ui/event_driven_ui.py
@@ -67,9 +67,12 @@ class EventDrivenUI(SimpleUI):
 
     @property
     def input_queue(self) -> asyncio.Queue[str]:
-        """Public accessor for input queue (backward compatibility).
+        """The queue incoming messages land on.
 
-        Prefer using handle_incoming_message() for routing messages.
+        Not a back-compat shim, despite how it was labelled: it is the public
+        read seam for the queue, and without it a caller (or a test) asserting
+        on queue state has to reach for the private attribute. Prefer
+        `handle_incoming_message()` for *routing* a message in.
         """
         return self._input_queue
